backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
aiomysql>=0.2.0
python-dotenv>=1.0.0
